make flashcard component reusable
