+++ src/cms/preview-templates/IndexPagePreview.js
@@ -13,7 +13,7 @@ const IndexPagePreview = ({ entry, getAsset }) => {
         heading={data.heading}
         subheading={data.subheading}
         description={data.description}
-        intro={data.intro || { blurbs: [] }}
+        intro={data.intro || { flashcards: [] }}
         mainpitch={data.mainpitch || {}}
       />
     )
